:truck: Move Graph into ocelot module
