+++ src/frontend/modules/ocelot/components/Graph/nodes/CircleNode.tsx
