+++ src/frontend/modules/ocelot/components/Graph/nodes/RectangleNode.tsx
